dailybuild:　Solution/Project files for Visual Studio 2013, and some obvious warning eliminations.
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@7444 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Creature/CRReachController.cpp
@@ -135,7 +135,7 @@ void CRReachController::Step() {
 				}
 			}
 			// 複数あって一部しか使わない場合
-			if (ikEffs.size() > 1 && numUseHands > 0 && numUseHands < ikEffs.size()) {
+			if ((int) ikEffs.size() > 1 && numUseHands > 0 && numUseHands < (int) ikEffs.size()) {
 				double sholderdistance[] = {ikEffUseFlags.size()};
 				for (size_t i=0; i<ikEffUseFlags.size(); ++i) {
 					if (baseJoints[i]) {
